SCREAM: add overload to 'subfield' that accepts subfield units
This is handy for tracers: Q is created with 'nondimensional'
units (since it's just a collection of disparate fields), but
the individual tracers are likely created with meaningful units.

Besides, not keeping the original units of each individual tracer
can cause field identifier mismatch when setting fields in
the atmosphere process that needed them.