Simplify computing landIcePressure and vert grid
While we supported 3 options for initializing the ice draft (SSH)
and landIcePressure, we now only use the simplest one, which uses
the denisty in the top layer as a guess for the effective density
of the ocean.

This merge significantly simplifies and clarifies the process for
creating the vertical coordinate and later initializing
landIcePressure in several test cases.  The subroutine for creating
the vertical coordinate ocn_init_vertical_grid() has been moved to
ocn_init_vertical_grids, where it clearly belongs.