Rewrite masks for apply_ablation
The masks are now defined first by the dynamic cells corresponding to the mask
type (with value=1).  Then the mask is extended forward to encompass
non-dynamic floating cells adjacent (value=2).  A second pass is made to include
'stranded' non-dynamic floating cells that can occur in sharp corners of
margins.  The inclusion of non-dynamic floating cells is appropriate for
any of the margin types.