Fix bmb restart issue
This change fixes an issue where the temperature solver and basal mass
balance calculation was not BFB on a restart in some situations.

The situation where is showed up was the 1km Thwaites mesh with:
    config_thermal_solver = 'temperature'
    config_thermal_calculate_bmb = .true.
The problem was not reproducible in the 4km spun-up MISMIP+ domain using
a similar configuration.  It's also not clear why this change solves the
problem.  But it's a relatively painless solution, so I'm committing it.