Removes shortwave energy conservation fix by default
Coupled testing showed unexpected extremely strong sensitivity to the
previous shortwave radiation energy fixer.  In that fix the excess
energy was used to heat the bottom layer.  This yielded a strong
response in tropical precipitation, dramatically increasing the double
ITCZ bias.  This removes that energy fixer and only tracks the solar
radiation that is discarded for extremely shallow cells