Add option to export gustiness from atmosphere.
This commit adds the variable ATM_SUPPLIES_GUSTINESS to env_run.xml. If
this variable is set to TRUE, the atmosphere exports an extra gustiness
value that is used by other components to calculate fluxes, rather than
scaling ubot and vbot to approximate the effect of gustiness. If this
variable is set to FALSE, there is no change from the current behavior.