SCREAM: fixing dynamics initialization sequence
Unfortunately, we cannot initialize dynamics during the initialize_impl
call, since the atm inputs have been read on the ref grid only (which
is likely NOT the dyn grid). Therefore, we complete the init sequence
during the first call to run_impl.

Also, added 'ps' as an input to dyn, which uses it to initialize dp3d.
It is only used at the 1st time step, but this is the only way to
read it in from file.