SCREAM: upgrade to machine specs for weaver
Weaver has two queues, both with V100 GPUs. Since sometimes the default queue
is full, it might be handy to submit to the other queue. So I added some bash
commands to run when the MACHINE_METADATA dict is built, which help to select
the second queue if the first one is full.