Improvements to implicit surface flux calculation.
Instead of simply assuming that the stress will change the lowest-level
wind, in this commit the land model takes an estimate of an
"equilibrium" stress compatible with the current wind (called
"tau_est"). Changes to the stress are then calculated assuming that the
*difference* between the new stress and this equilibrium value is what
affects the lowest-level wind. Currently, the estimation of the
equilibrium stress is simply the stress that was applied by CLUBB in the
previous time step, as this is both easy and leads to a quite stable
method.

This commit also removes the u_diff/v_diff fields added in the previous
commit.