SCREAM: do not fetch during test-all-scream run.
Instead, check if $baseline_ref in the current repo is ahead of
the sha stored in the baseline dir. If ahead, mark baselines as
expired. However, if $baseline_ref in the current repo is *behind*
the sha stored in the baseline dir, we crap out, informing the
user that he should fetch a newer version of the repo.