Add switches to control implicit stress method.
The associated CIME commit has added the ATM_FLUX_INTEGRATION_METHOD
variable to env_run.xml. If this option is set to "explicit", the
original flux calculations are used. If this option is set to
"implicit_stress", the new implicit stress code is used.

The explicit (default) setting is bit-for-bit with the version of E3SM
master that this code (and associated E3SM commit) was branched from.

This commit also required changing a number of new "optional" arguments
to CLUBB to be pointers instead. This is because the Intel compiler did
not support using null pointers or unallocated arrays to represent
optional arguments that are not present (until very recently).