Initialize standard name when reading restarts
Initialize standard name for fields in the history list while
reading data from restart files. A similar action has been
performed for mixing ratio before.

Without this initialization the string can contain invalid values
and result in the value (uninitialized/invalid value) being
written out as field attributes for the variables in the EAM
history files. For GNU, the string might be empty in certain
procs and non-empty in others, such that pio_put_att is not
called collectively on it (simulation either hangs or crashes).