SCREAM: improved ColumnOps impl, to work with non-packed types
The previous impl was templating the class over pack information,
and the impl of the methods contained both pack_size 1 and >1 in
the same method (using a compile time if to switch).
This prevented instantiating the function for a non-pack type,
even if the pack_size 1 branch did not contain pack-specific code.

To fix this, remove pack specs from template signature of the class,
and use SFINAE to select impl of the methods depending on pack size.