Comment unused basalOceanHeatflx call that currently causes error in glc log
This variable is a placeholder for a future addition in MALI that has
not been implemented yet.  The assignment is already commented, but the
attempt to get it from an MPAS pool was causing harmless but confusing error messages
to be printed in the GLC log.  The pool call is now commented.