Add error checks for required number of halos (3)
FO velocity solver requires 3 halos.
edgeMask calculation also requires it.
Checks have been put in both routines so that if one of those
restrictions is ever relaxed, the other will still be triggered.
A previous commit changed the default num_halos but did not ensure the
runtime namelist was actually set that way.