Response to reviewer comments, and some code cleanup.
This commit moves the common physics tests out of the src/physics/share/tests and
into src/share/tests which reflects the location of the common_physics code.

In response to reviewer comments the inline comments have been moved around so that the
header to include, scream_common_physics.hpp, has the more descriptive explanation of
what the function does, while the impl has a smaller description.

This commit also renames the common physics functions to move away from using `get_` and
either using `calculate_` when appropriate, or in the case of exner, just `exner_function`.