Allow MMF tests to build unthreaded on Chrysalis
OpenMP offload directives in the new CRM code will not build as-is with intel on Chrysalis. Since MMF in general will not run threaded yet, and because we do not intend to run the OpenMP CRM on Chrysalis with the intel compiler, allow building this code without OpenMP flags so that nightly tests will still build and run. Note that MMF tests were always unthreaded, but there was a special test mod added to build with OpenMP flags even if NTHRDS was set to 1 on Chrysalis only. This commit simply disables forcing the MMF tests to run with OpenMP flags.