Modifications so Cloud-J will work with OpenMP properly
The initialization routines Cloud-J were stopping the code from
running with OpenMP. Pat Worley provided a temporary fix with OpenMP
directives. However, this commit creates a separate initialization
routine, and moves and is now called from within chem_init.  This also
means it is only called once, which is the obvious long-term solution.

The initialization is still read separately by each processor.
Fortunately, the files are small and this only needs to be done once.

[BFB]