Minor updates for implicit_stress option for UrbanFluxesMod.
This commit moves an implicit_stress check outside of a loop in
UrbanFluxesMod so that the loop is skipped entirely if that option is
off.

This commit also fixes a typo in the initial setting of tau_diff that
could slow convergence of tau iterations for the implicit stress method
(in practice the previous value should still be large enough that this
should rarely/never have a significant effect on results).