Big fix to the conversion between mmr and vmr.
The formula was incorrect and did not take into account that we are
working with wet-mass-mixing-ratios in scream.

This commit fixes the conversion functions and the tests associated with them.

We also had to back out the wet mass mixing ratio for qv from the volume
mass mixing ratio for qv in the rrtmgp_stand_alone test.