Replace active gases lookup with case insensitive strings via EKAT.
This commit provides the flexibility to define active gases in a run
with any combination of cases, for example ch4, Ch4 and CH4 should all
be valid choices.

This is an important functionality for the eventual coupling with
prognostic aerosols where individual aerosol packages may have adopted
different standards for defining gas names.