Homme: Work around or fix an issue re: threading in standalone Homme.
It was reported that some standalone Homme SL transport tests were crashing with
multiple threads. The failure is an inconsistent shared-pointer count in a
Kokkos array.

This error has not appeared on, e.g., Cori-KNL, despite always running
threaded there. In addition, I'm not seeing a bug in the SL C++ code. But I
do know how to avoid the problem, and this commit does that.

[BFB]