SCREAM: fix to AD logic related to group requests
In particular, related to handling group requests with a 'relative' request
on a different grid. More specifically, if request gr1 on grid A has a
relative gr2 on grid B, we ensure that an equivalent of gr2 on grid A
is registered in the FM on grid A. This fix was mostly dictated by
dynamics registering the group tracers_prev on the dyn grid, but not
registering the tracers group on the dyn grid, which ultimately caused
a crash when the FM was trying to ensure qv is the first tracer.