Force a read of 'forcing' streams on init
Previously, input files used as time-varying forcing failed to be read
at the initial model time.  This commit forces input streams with an
input_interval set to a time duration to be forced to be read at the
model initial time.  They are read in the same we these "forcing" files
are read during timestepping by using the most recent previous time
level from the input file.