Ensure sync is called before freedecomp
Moving the freeing of decompositions before the file close.

Without the change when freeing io descriptors Scorpio performs
a syncfile on the File handle (that is no longer valid because
it was closed) resulting in an abort (sync fails since it
cannot find the associated file handle).

(This change was made in E3SM master in PR #3607 but was not
merged to the maint-1.0 branch)