Introduce a new atmpsphere process, spa
This commit introduces the atmosphere process SPA and establishes
the first step in the SPA main set of calls, vertical interpolation.

At the moment the SPA interface is a bit klunky and will require a bit
of work to get full functionality.

For example, the prescribed aerosol values are only CCN3 at this point, and
they are held fixed in time.

Similarly, the CCN3 values are assumed to be on the same mesh as the simulation.

In summary, only vertical interpolation is done at this point.  The next steps will
be to bring in
1) temporal interpolation - allowing us to have all 12 months of data.
2) and horizontal interpolation - allowing us to use coarse forcing data for fine
   resolution simulations.
3) Expand the list of aerosols to include all aerosols, not just CCN3