SCREAM: allow using raw data instead of Field's in AtmosphereInput
The user can now init the input class in two ways:
* Pass 1d host views and layouts of fields to read
* Pass a FieldManager

In the former case, we *assume* the views are contiguous,
that is, no padding or strides. In the second case, we
can handle strided views (e.g., if fields are subfields)
or padded views, and we will make sure data is synced to dev.