Add field property checks to specific fields in specific processes.
Prior to this commit, the only field property check that was on by
default was checking for NaN.  Although, it makes sense that as
processes register a field it might also be possible to add specific
field property checks that make sense in the context for that process.

For example, putting a bounds check on temperature, or making sure
mixing ratios are always positive.