Add timers to EAM to provide context to SCORPIO timers
SCORPIO has introduced some new timers. Unfortunately, the existing
timers in EAM are not sufficient to localize where the SCORPIO calls
are coming from. Here new timers are added to provide the missing
context. A few timers are also added to provide context for these
added timers and for some of the already existing timers.

The logic for the timer physpkg_st1 is also changed, so that the
count during the initialization is '1' instead of '2', and the count
is halved in the run loop.

Also, some places have timer names in double quotes in calls to
t_startf/t_stopf and some have them in single quotes. Changed double
quotes to single quotes in some files, to impose consistency within
the file.

Finally, a timer name typo (prim_run_sybcycle) is corrected (to
prim_run_subcycle).

For an initial run, 28 new timers are introduced into the
initialization and 13 new timers are introduced into the run loop.
Comparing the number of lines in the process 0 timing file for
a watercycle case, there are 41 more lines of timer data after
implementing these modifications, in agreement with the new
timer count.

For a continuation run, 30 new timers are introduced into the
initialization and 13 new timers are introduced into the run loop.
Comparing the number of lines in the process 0 timing file for
a watercycle case, there are 43 more lines of timer data after
implementing these modifications, in agreement with the new
timer count.

BFB