Eliminate recursve call to the CPL:RUN timer
The following logic in cime_run_atmocn_setup (in cime_comp_mod.F90):

   call t_drvstartf ('CPL:ATMOCNP',cplrun=.true.,barrier=mpicom_CPLID,hashint=hashint(7))
   ...
   if (do_budgets) then
      call cime_run_calc_budgets3()
   endif
   ...
   call t_drvstopf  ('CPL:ATMOCNP',cplrun=.true.,hashint=hashint(7))

and in cime_run_calc_budgets (also in cime_comp_mod.F90):

   call t_drvstartf ('CPL:BUDGET0',cplrun=.true.,budget=.true.,barrier=mpicom_CPLID)
   ...
   call t_drvstopf ('CPL:BUDGET0',cplrun=.true.,budget=.true.)

nests calls to the CPL:RUN timer, leading to recursion being recorded
in the timing library. This is not a bug, but it is also not necessary.
Everything in the CPL:ATMOCNP timer is marked as being in the CPL:RUN
timer. It adds nothing to redundantly mark the CPL:BUDGET0 timer as
being in the CPL:RUN timer.

Here an optional logical parameter 'in_cplrun' is added to
cime_run_calc_budgets3, and to cime_run_calc_budgets1 and
cime_run_calc_budgets2 for consistency, such that if set to 'true' it
indicates that the call to the routine is already inside of the
CPL:RUN timer and that this should not be set again within the
routine.

BFB