Implement the cosine zenith angle calculation within the rrtmgp interface.
This commit removes cosine_zenith as a required variable for rrmgp, now
that we are able to calculate these values on the fly using lat, lon and
time.

This commit also establishes a lat/lon pair for the rrtmgp stand alone test
that will ensure that the resulting cosine_zenith angle will be 0.86, which
is what it was hard-coded to prior to this commit.