Separate drv_resume file and logical
Make drv_resume a logical and use it to determine if
a restart file should be read.  Introduce drv_resume_file and use it
where drv_resume was assumed to be a filename.

Also add a log message for the real restart read at the start of a run.

Fixes #4513