SCREAM: changes to mach specs
Increase make parallel resources of SNL machines to match the
number of hw threads. On the other hand, keep the ctest parallel
resources equal to the number of cores (or GPUs).
That's b/c hw threads can all be used for building, while for testing
it's best to bind one mpi rank per core (or GPU).