SCREAM: allow I/O on dynamics grid
More generally, allow I/O on arbitrary grids, provided that
they store a 'unique grid', and that the grids manager can
build a remapper from the grid to its unique grid.

The I/O classes will then perform I/O on a copy of the fields
on the unique grid, and then use the remapper for copy
to/from the original field manager variables.