Make simlar adjustments to the output routine in the scorpio interface so that integer output is possible.
Note, just as with the read_input there is no test for the integer functionality, yet.
Currently output uses the output manager which relies on the field manager for output.
The field manager assumes all data is real defined, so it will take more work to get
an integer type output/input test going.