SCREAM: in P3, use ETI also if relocatable device code is used on CUDA
Relocatable device code (RDC) is less performant than a single
translation unit (TU), but it is significantly faster to build
when one has several cores (as it is usually the case), since
each TU is smaller and built in parallel.

The files atmosphere_microphysics.cpp and p3_functions_f90.cpp
are very long to build, since they contain, directly or indirectly,
a call to virtually all the functions in p3_functions.hpp, which,
witout RDC, have to be compiled when building those files.
In a debug build (-O0), each of these two files takes about
5.5 minutes to build, counting all steps (nvcc, host compiler, and
assembler). With RDC on, they build in about 30 seconds.