SCREAM: fix required FieldGroup logic in AtmosphereProcessGroup
In sequential splitting, an APG can have a FieldGroup (FG) or a
single Field that is listed as requried by one of its APs, but
is computed by one (or more) of previous APs in the group.
In that case, the APG should *not* expose the FG (or the Field)
as 'required' by the APG.

The APG class cannot detect if a FG required by one of the
group's APs is indeed an input to the APG or not while still
in the 'set_grids' call. In fact, at that time, the APG has no way
to know the names of the fields in the group. Some fields might
be added to the FG by an AP that is not in this APG. Even for
a single Field this is not possible, for it might belong to
a group computed by a previous AP.

The APG can detect if a FG or Field is indeed required by the APG
as a whole only when FG's and Field's are set in the APG via
calls to set_required_[field|group], *assuming* that the
computed FG's and Field's are set first.