Introduce a function to SPA to update the timestate.
The new update_spa_timestate function will update the current time stamp
which is used for temporal interpolation in spa_main.  It will also check
if we have switched to a new month.  If so, it will update the data structures
for the beggining and end of the month.

Also a few minor fixes for bugs left over from the last commit.