Adjust the remap weights function such that it supports runs with multiple mpi ranks.
This commit added the degrees-of-freedom view as an argument to the
remap weights function such that instead of storing all of the remap
weights, only those weights that correspond to a global column id that
is handled by this particular rank will be stored.

The spa unit test has been updated to check for multi-rank jobs, including
an update to the data file itself to have more columns and the test setup, to
create a test for each MPI rank defined in SCREAM testing by the maximum MPI
ranks for testing.