Moves the ELM tools to one directory up
At the begining of E3SM, the land model had `clm4_0` and
`clm4_5` models and the common tools were in the `shared`
directory. Since, `clm4_0` has been deleted and tools
in the `clm4_5` and `shared` are consolidated in a single
directory.

The relative path to `queryDefaultNamelist.pl` is also
updated