Big refactor to taskset handling in test-all-scream
The prior implementation assumed an unrestricted test-all-scream process
with access to the entire machine. lassen breaks this assumption because
compute node have logical cores 0-8 reserved. This required a new implementation
that did not make that assumption.

We now use the python tpl "psutil" to get a list of all logical cores
available to test-all-scream and we use that list to manage tasksetting.

With the availability of psutil, we no longer need to hardcode resource
counts in the MACHINE_METADATA dict. We can accurately probe for what's
available.

With these changes, we can safely limit test-all-scream to any (contiguous) subset
of the machine's cores and things should work.