SCREAM: bug fix in scorpio F90 interface
When opening a file for read, we need to set the file purpose in the
pio_file object *before* calling eam_pio_openfile, since it is later
used as open mode in the netcdf routine. For some reason, we were
not hitting this bug anywhere, until the spa new interpolation
came in, but we *were* using uninited memory.