Homme: In FV->GLL remap, fix input handling of q for ftype 0.
The previous implementation was based on a misunderstanding of the input Q type
when ftype is 0. I had thought both the input (Q) and the output (FQ) were
tendencies when ftype=0. But in fact neither is, just like when ftype > 0. In
other words, gllfvremap_mod is independent of ftype.

Also add another FV<->GLL API test to the C++ unit test.