Explicitly use pool allocator for ESMT test
Make sure we use the pool allocator for the ESMT test. This used to be
the default in YAKL, but commit fcf7ba19 appears to have changed that
behavior, and for some reason that is causing non-BFB behavior with
this test. We do not understand why, but reverting this behavior should
make this test BFB with master, and get the ESMT test to pass again on
Chrysalis.