Make changes that were suggested in #1289 but didn't get fixed before the PR was automerged.
This commit changes the way the filenames are handled in the scorpio_output
class to make it easier to manipulate them given simulation information.

Orginally the time stamp was the only simulation data added to the filename.
This commit adds the option to also add the number of mpi ranks that were
used for the simulation, and creates a framework for adding other options in
the future if needed.

There are also a few cosmetic changes that should have been made in 1289 but
weren't.