Hommexx: fix initialization of tracers forcing
I am not sure if/when we hit the specific code branch
that had the wrong initialization. Still, it was buggy,
and now it's fixed.

Namely, the buggy initialization was passing only num_elems
to the ctor of the view, but fq has 2 dyn dimension, the latter
begin the number of tracers (qsize).