Fix how the iodesc tags are created in the scream_scorpio interface to allow for variables with the same basic layout but on different grids.
For example, if loading data from two different files defined on two
different grids, but with the same layout, both would be assigned the
same pio_iodesc structure.  But the iodesc structure is used to determine
how big the dataset is too.

So if loading variable A from file f0 and f1 where A has dimensions COL,LEV.
But on f0 variable A has 10 columns and 4 levels while on f1 it has 20 columns
and 6 levels, in the past the IO would assume that in both cases it was reading
data with 10 columns and 4 levels.  No it will not only distinguish between layouts
(i.e. COL,LEV) but also length (i.e. 10,4 vs. 20,6).