Expanding the OpenACC region in the split time integration routine
Expanding the OpenACC region in the split time integration routine and
partially in the other time integration routines. As part of this,
added routines to get scalars instead of pointers to make it
easier to port to OpenACC. These include:
   mpas_get_dimension_scalar
   mpas_get_config_scalar
   included duplicate function varinp_xxx as well as config_xxx for
     config variables to get scalar versions and added to gen_inc.c
All of the framework changes were supplemental features and the existing
functionality was not touched to avoid breakage
Also added gpu version of mpas_vector_reconstruct function to be used
by split time integration until all the uses can be done on the gpu.