Replace actualWhen with saveActualWhen
This commit modified the previous commit by removing the actualWhen
argument to the MPAS_stream_mgr_read() routine and replacing it with an
optional argument saveActualWhen.  The actualWhen argument was ambiguous
when MPAS_stream_mgr_read was called in a way such that multiple streams
were read - the actualWhen value returned came from the final stream
read and could possibly be wrong for the other streams that were read.
Given the previous commit added the mostRecentAccessTime attribute to
the MPAS_stream_list_type type, the actualTime argument was somewhat
redundant anyway.

Now, instead of being able to have the actualTime returned as an
argument, there is the option to have the actual time saved to the new
attribute.  It can then be retrieved from the stream cursor as needed.

The saveActualWhen argument is optional because some streams do not
contain xtime and trying to obtain ‘actual when’ in those cases will
cause the reading of the stream to fail.  Therefore, it is important to
be able to not seek this information if possible.  If the saveActualWhen
argument is not provided to MPAS_stream_mgr_read(), its value defaults
to false.

This commit also changes the land ice core to stop using actualWhen and
replace it with saveActualWhen.  Neither the ocean nor seaice cores use
actualWhen so no other changes are necessary.