Update the capabilities of SCREAM's scream_input.yaml file
We add some customized syntax to increase the expressiveness
of the items in this file:

'${VAR}' will be used to refer to env variables in the CIME case

'<cond : key1 => val1 : key2 => val2 : elseval>' will be used to express conditional
statements. If cond matches key1, then the expression evaluates to val1; if cond
matches key2, then the expression evaluates to val2; if it matches neither, then
the expression evaluates to elseval. The elseval component of this expression is optional.

This should give SCREAM the ability to express what it needs in this file
without having to resort to lots of hardcoded special handling.