SCREAM: allow storing also CG dofs in a SE grid
SE grid are considered to have nelem*ngp*ngp independent dofs.
This corresponds to a Discontinuous Galerkin (DG) way of
discretizing a field. In order to perform PD remap, however,
we need knowledge of how matching edge dofs across bordering
elements are related. To this end, we store also a view with
the numeration of dofs in a Continuous Galerkin (CG) fashion.