SCREAM: fix in PhysicsDynamicsRemapper
When computing loops sizes, do not rely on either physics or dynamics
views dimensions, since one might be packed, but the other might not.
To fix this, store the number of levels of each field, and compute
the last dim extent (for 3d fields) based on that and the pack size.

This bug was in the local remap fwd 3d. The bwd remap was implemented
correctly, but had the column idx striding faster than the level idx,
which causes poor memory access on GPU, so I propagated the fix also
to the bwd remap, to use better memory access.