Fix filename interval for 2 analysis member restarts
The Eliassen-Palm and time-filters analysis members both had
hard-coded restart file intervals of 1 year.  This causes restarts
with these analysis members to fail in situations where the restart
interval does not match the file interval.  This merge changes the
filename interval to be the same as the output interval to match
restart streams in all other analysis members.