Reset default cosp_ncolumns to 10 for ne1024
The hardwired default for cosp_ncolumns is 50. This PR resets it to
10 in namelist_defaults_eam.xml for ne1024 grid. When cosp_lite is true,
cosp_ncolumns is always set to 10. cosp_ncolumns is also set to 10 when
the any of the following is true: cosp_passive, cosp_active, cosp_isccp,
and cosp_amwg. Note that setting cosp_lisccp_sim = .true. does not lead to
cosp_isccp = .true. during runtime. But it does serve to set lisccp_sim to
true, which is the actual control for everything regarding ISCCP simulator

[BFB] no impact when not using ne1024 grid