SCREAM: minor change in OutputManager regarding restarts
When looking for restart filename in rpointer.atm, if timestamps are
present in the filenames, only consider timestamps that are in the
past. This is helpful when debugging, and re-running the same restart several
times, without the need of resetting the rpointer file.

Note: to this end, OutputManager needs to store both the simulation
      start time (the case t0) as well as the run start time.