Work-around an FPE in drydep calculations
A recent bugfix in some code in the land model results in a floating
point exception when running SCREAM with debug flags. This occurs from
the combination of using prescribed aerosol with setting dry-deposition
options, which should not be used in the first place with prescribed
aerosol. We had mistakenly carried the dry-deposition namelist option
over from the EAM F2010 use case file to our SCREAM use case files, so
this was triggering the calculation, that ultimately ends up not even
being used. So removing this namelist option from the use case files
results in a BFB simulation, although the test does throw a DIFF because
the field list in the cpl.hi files differ (because we are no longer
doing this unused calculation, which apparently needed some extra
variables in the cpl.hi files).