Enable iac data pass through coupler to land
The iac data is successfully passed to the land model through the
coupler. The grid mapping is not correct yet, so the land model
fails due to incomplete pft sums. As a result, the actual updating of
the land model values is currently disabled. Some diagnostic code has
been cleaned up. This requires the giac branch aldivi/shippert/merge to
work.