SCREAM: add a field property check wrapper
The wrapper can store two FPC pointers for the check
and repair phase, allowing two different criteria for
the two phases. E.g., dynamics needs to check that
tracers are not negative, and then clip them to 0
from below. However, very tiny negative values can occur,
and should not trigger a failure. With this class, we can
create a LowerBound check with small (in magnitude) negative
threshold for the check phase, and a LowerBound check
with 0 threshold for the repair phase.