SCREAM: minor mods to BuildCprnc
While we usually do not build scream tests in CIME builds (and therefore
not need cprnc in CIME builds), it is technically possible to build
them. If that's the case, the BuildCprnc script was buggy, in that it
assumed that the current project was scream.

This commit fixes that, and uses established paths to deduce cprnc
source folder location.