SCREAM: added level of indirection for Field deep_copy method
This allows to do something like f.deep_copy(0) even for real fields,
without using the type of the scalar to deduce the data type needed
to extract the internal view (which in this case would be int).

The input value must be convertible to the field stored data_type.