Add testmod for prognostic volcanic aerosol config
Add testmod (eam-prognostic_volcaero) for prognostic volcanic aerosol
configuration to be used with F20TR compsets. This is primarily to
support testing CLDERA simulation configurations with regular nightly
(or at least integration) testing. Note that the user_nl_eam namelist
for the testmod does use two input files that are not in the standard
e3sm input data repository and location will be machine-dependent for
now. At the moment, it is hard-coded for the location on mappy. We
should think about moving these input files to the central e3sm
inputdata repository to make this configuration more flexible.