Add e3sm_cldera test suite
Add e3sm_cldera test suite to enable running a small set of regression
tests specific to CLDERA development. At present, the suite includes
just F20TR configurations, one with the new v2 default configuration
that uses MPASSI instead of CICE and thus requires a different grid
configuration, and one that forces the use of CICE to be backward
compatible with ongoing simulations. For better test coverage, one of
these is a restart test, and one is a smoke test with debug flags.