Refactor handling of scream_input.yaml
A fair bit of effort was made to preserve the comments in the repo
file for screan_input.yaml during the processing of that file for
the case. The idea was that users would be interacting with this
file in their cases and we wanted them to have access to our comments
directly to assist them.

YAML does not have any internal mechasisms for preserving comments
when parsing/dumping YAML content, so I had added some hacky code
to read comments separately and reestablish them in the file manually.

After a few months of developers creating CIME SCREAM V1
cases, it seems this approach has been a source of syntax errors
in this case's version of this file, so this PR is a rethink of
the approach.

Instead of relying solely on inline comments in the repo's scream_input.yaml,
we provide a separate SCREAM_INPUT_README file for full documentation
on how to use scream_input.yaml.