Enable setting fixed orbital parameters
Enable setting fixed orbital parameters. This is needed for aquaplanet
configurations, in which we need to set orbital parameters to perpetual
equinox conditions (by setting eccentricity and obliquity to zero). This
exposes these parameters to the AD through YAML options, and adds
default values to the scream/data/scream_input.yaml file based on
compset regex. The new behavior now checks if non-negative orbital
parameters are passed via the yaml file, and if they are, bypasses the
computation of orbital parameters based on year in favor of using the
fixed parameters. Note that ALL parameters must be non-negative, or else
all parameters will be computed instead in shr_orb_params.