Establish a atm. proc subclass for diagnostics
This commit establishes the framework for a new atmosphere_diagnostics
class which is a sub-class of the atm_process class.

A diagnostic class is different than an atmospehre process primarily in
that a diagnostic will not be allowed to alter any fields in the field
manager.  i.e. no "computed" or "updated" fields.

Instead, the diagnostic has the unique trait of producing an output
field directly with the diagnostic values in it.  This can then be
used for model output or field property checks.

This commit also adds a new test to the atm_proc tests checking
diagnostics.