Add halo update on calvingVelocity
Testing indicates that a halo update is needed either on submergedArea
or on calvingVelocity before calling li_apply_front_ablation_velocity.
Those two halo updates give the same result, so updating calvingVelocity
because submergedArea is a local variable. Alternatively, submergedArea
could be added to Registry, but this is simpler. Behavior is still
different across different numbers of processors, even without velocity
solver or advection. This is likely due to an issue within li_apply_front_ablation_velocity,
and will be addressed in another PR.